requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2
django-filter==24.2
django-ranged-response==0.2.0
django-recaptcha==4.0.0
django-robots==6.1
django-simple-captcha==0.6.0
djangorestframework==3.15.2
drf-yasg==1.21.7
idna==3.7
importlib_resources==6.4.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Markdown==3.6
packaging==24.1
pillow==10.3.0
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
setuptools==70.3.0
sqlparse==0.5.0
swagger-spec-validator==3.0.4
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
wheel==0.43.0
